chore: Split tests in success and error
