+++ crates/miden-block-prover/src/tests/mod.rs
@@ -1,2 +1,4 @@
-mod proposed_block;
+mod proposed_block_errors;
+mod proposed_block_success;
+
 pub(crate) mod utils;
