+++ crates/miden-block-prover/src/tests/proposed_block_errors.rs
@@ -1,6 +1,5 @@
 use std::{collections::BTreeMap, vec::Vec};
 
-use anyhow::Context;
 use miden_lib::transaction::TransactionKernel;
 use miden_objects::{
     account::AccountId,
@@ -19,49 +18,6 @@ use crate::tests::utils::{
     generate_note_with_asset, generate_tx, generate_tx_with_unauthenticated_notes,
     generate_untracked_note, setup_chain, setup_chain_with_auth, ProvenTransactionExt, TestSetup,
 };
-
-#[test]
-fn proposed_block_success() -> anyhow::Result<()> {
-    let TestSetup { mut chain, mut accounts, mut txs, .. } = setup_chain(2);
-    let account0 = accounts.remove(&0).unwrap();
-    let account1 = accounts.remove(&1).unwrap();
-    let proven_tx0 = txs.remove(&0).unwrap();
-    let proven_tx1 = txs.remove(&1).unwrap();
-
-    let batch0 = chain
-        .propose_transaction_batch([proven_tx0.clone()])
-        .map(|batch| chain.prove_transaction_batch(batch))
-        .context("failed to propose transaction batch")?;
-
-    let batch1 = chain
-        .propose_transaction_batch([proven_tx1.clone()])
-        .map(|batch| chain.prove_transaction_batch(batch))
-        .context("failed to propose transaction batch")?;
-
-    let batches = [batch0, batch1];
-    let block_inputs = chain.get_block_inputs(&batches);
-
-    let proposed_block = ProposedBlock::new(block_inputs.clone(), batches.to_vec()).unwrap();
-
-    assert_eq!(proposed_block.batches(), batches);
-    assert_eq!(proposed_block.block_num(), block_inputs.prev_block_header().block_num() + 1);
-    let updated_accounts =
-        proposed_block.updated_accounts().iter().cloned().collect::<BTreeMap<_, _>>();
-
-    assert_eq!(updated_accounts.len(), 2);
-    assert_eq!(updated_accounts[&account0.id()].transactions(), &[proven_tx0.id()]);
-    assert_eq!(updated_accounts[&account1.id()].transactions(), &[proven_tx1.id()]);
-    assert_eq!(
-        updated_accounts[&account0.id()].final_state_commitment(),
-        proven_tx0.account_update().final_state_hash()
-    );
-    assert_eq!(
-        updated_accounts[&account1.id()].final_state_commitment(),
-        proven_tx1.account_update().final_state_hash()
-    );
-
-    Ok(())
-}
 
 #[test]
 fn proposed_block_fails_on_empty_batches() -> anyhow::Result<()> {
